Windows fixes
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@9570 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/LOSTexture.cpp
@@ -156,7 +156,7 @@ void CLOSTexture::RecomputeTexture(int unit)
 	if (m_Texture)
 	{
 		CmpPtr<ICmpTerrain> cmpTerrain(m_Simulation, SYSTEM_ENTITY);
-		if (!cmpTerrain.null() && m_MapSize != cmpTerrain->GetVerticesPerSide())
+		if (!cmpTerrain.null() && m_MapSize != (ssize_t)cmpTerrain->GetVerticesPerSide())
 			DeleteTexture();
 	}
